By default initialize the session with UI_Console
