+++ lib/debug.rb
@@ -1,1 +1,2 @@
 require_relative 'debug/session'
+require_relative 'debug/console'
